+++ tests/Pest.php
@@ -4,3 +4,4 @@ use D076\Tracing\Tests\TestCase;
 
 uses(TestCase::class)->in('Feature', 'Unit');
 uses(D076\Tracing\Tests\CustomConnectionTestCase::class)->in('Integration');
+uses(D076\Tracing\Tests\DisabledTestCase::class)->in('MasterSwitch');